+++ db/migrate/20140706182925_add_stop_after_to_schedule.rb
@@ -1,0 +1,6 @@
+class AddStopAfterToSchedule < ActiveRecord::Migration
+  def change
+    add_column :schedules, :stop_after, :integer
+    add_column :schedules, :stop_suspend, :boolean
+  end
+end